django_frontend/requirements.txt
Django==3.0.2
djangorestframework==3.11.0
django-rest-auth==0.9.5
django-rest-knox==4.1.0
requests==2.22.0
psycopg2==2.8.4
google-python-cloud-debugger